+++ src/services.rs
@@ -7,6 +7,7 @@ use crate::Result;
 use crossbeam_channel::{Receiver, Sender};
 use std::sync::{Arc, Mutex};
 
+pub mod automator;
 pub mod buienradar;
 pub mod clock;
 pub mod db;
@@ -40,5 +41,6 @@ pub fn new(service_id: &str, settings: &ServiceSettings) -> Result<Box<dyn Servi
         ServiceSettings::Db(settings) => Ok(Box::new(db::Db::new(service_id, settings))),
         ServiceSettings::Buienradar(settings) => Ok(Box::new(buienradar::Buienradar::new(service_id, settings)?)),
         ServiceSettings::Nest(settings) => Ok(Box::new(nest::Nest::new(service_id, settings))),
+        ServiceSettings::Automator(settings) => Ok(Box::new(automator::Automator::new(service_id, settings))),
     }
 }